refactor: assetId 가 Integer 이었던부분 Long으로 변경

+++ common-mysql/src/main/java/com/stocat/common/domain/order/Order.java
@@ -38,7 +38,7 @@ public class Order extends BaseEntity {
     private Long memberId;
 
     @Column(nullable = false)
-    private Integer assetId;
+    private Long assetId;
 
 //    @Column(nullable = false)
 //    private AssetsCategory category;
