+++ trade-api/src/main/java/com/stocat/tradeapi/order/service/dto/OrderDto.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public record OrderDto(
         Long id,
         Long memberId,
-        Integer assetId,
+        Long assetId,
         TradeSide side,
         OrderType type,
         OrderStatus status,
